+++ src/backend/AxiosRequestBackend.ts
@@ -1,9 +1,7 @@
-import RequestBackend, {RequestBackendErrorInfo, RequestOperation} from "./RequestBackend";
-import {ApiResponse}                                               from "../ApiTypes";
+import RequestBackend, {ConvertedApiResponse, RequestBackendErrorInfo, RequestOperation} from "./RequestBackend";
+import {ApiResponse}                                                                     from "../ApiTypes";
 import type {AxiosError, AxiosResponse, AxiosStatic}               from "axios";
 import RequestContext                                              from "../RequestContext";
-import {inferResponseType}                                         from "../ApiUtils";
-import {convertToRequestError, RequestErrorCode}                   from "../RequestError";
 
 let axios: AxiosStatic;
 
@@ -29,17 +27,8 @@ export default class AxiosRequestBackend implements RequestBackend<AxiosResponse
     return undefined;
   }
 
-  async convertResponse<T>(context: RequestContext, response: AxiosResponse): Promise<ApiResponse<T>> {
-    const contentType = response.headers["content-type"];
-    const inferredResponseType = inferResponseType(contentType);
-    // expand to array buffer once we support that in inferResponseType
-    if (inferredResponseType === "text" && context.responseType === "json") {
-      throw convertToRequestError({
-        error: new Error(`[api-def] Expected '${context.responseType}' response, got '${inferredResponseType}' (from 'Content-Type' of '${contentType}')`),
-        code : RequestErrorCode.REQUEST_MISMATCH_RESPONSE_TYPE,
-      });
-    }
-
+  async convertResponse<T>(context: RequestContext, response: AxiosResponse): Promise<ConvertedApiResponse<T>> {
+    (response as ConvertedApiResponse<T>).__lowercaseHeaders = (response as any)._lowerCaseResponseHeaders;
     return response;
   }