+++ src/backend/RequestBackend.ts
@@ -10,12 +10,14 @@ export interface RequestBackendErrorInfo {
   code: string;
 }
 
+export type ConvertedApiResponse<T> = ApiResponse<T> & {__lowercaseHeaders?: any};
+
 export default interface RequestBackend<R = any> {
   readonly id: string;
 
   makeRequest(context: RequestContext): RequestOperation<R>;
 
-  convertResponse<T>(context: RequestContext, response: R, error?: boolean): Promise<ApiResponse<T>>;
+  convertResponse<T>(context: RequestContext, response: R): Promise<ConvertedApiResponse<T>>;
 
   extractResponseFromError(
     error: Error,