+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - add `ResponseOf`, `QueryOf`, `BodyOf`, `ParamsOf` to quickly get types from endpoints
 - new `queryParser` config option to allow custom query string parsing
+- better errors
+- automatically lowercase header names
 
 ## Fixes